Update stock quantity field when a product is added/removed from a cart

+++ server/controllers/cart.js
@@ -47,24 +47,43 @@ const addItem = async (req, res) => {
 
   if (userId) {
     try {
+      // get the cart based on the users id
       const cart = await pool.query("SELECT * FROM carts WHERE user_id = $1", [
         userId,
       ]);
       const cartId = cart.rows[0].id;
 
-      await pool.query(
-        `INSERT INTO cart_items(cart_id, product_id, quantity)
+      // Check if there's enough stock for the product
+      const product = await pool.query(
+        "SELECT stock_quantity FROM products WHERE id = $1",
+        [product_id]
+      );
+      const availableStock = product.rows[0].stock_quantity;
+
+      if (availableStock >= quantity) {
+        // decrease stock quantity from products table
+        await pool.query(
+          "UPDATE products SET stock_quantity = (stock_quantity - $1) WHERE id = $2",
+          [quantity, product_id]
+        );
+        // insert or update product into the user's cart
+        await pool.query(
+          `INSERT INTO cart_items(cart_id, product_id, quantity)
         VALUES($1, $2, $3) ON CONFLICT (cart_id, product_id)
       DO UPDATE SET quantity = cart_items.quantity + $3 RETURNING *`,
-        [cartId, product_id, quantity]
-      );
+          [cartId, product_id, quantity]
+        );
 
-      const results = await pool.query(
-        "SELECT p.*, c.quantity, ROUND((p.price * c.quantity):: NUMERIC, 2) AS subtotal FROM cart_items c JOIN products p ON c.product_id = p.id WHERE c.cart_id = $1",
-        [cartId]
-      );
+        const results = await pool.query(
+          "SELECT p.*, c.quantity, ROUND((p.price * c.quantity):: NUMERIC, 2) AS subtotal FROM cart_items c JOIN products p ON c.product_id = p.id WHERE c.cart_id = $1",
+          [cartId]
+        );
 
-      return res.status(200).json({ cart: results.rows });
+        return res.status(200).json({ cart: results.rows });
+      } else {
+        // Not enough stock, return an err
+        return res.status(400).json({ error: "Not enough stock available." });
+      }
     } catch (error) {
       console.error(error.message);
       return res.json({ error: error.message });
@@ -82,6 +101,12 @@ const decreaseItem = async (req, res) => {
         userId,
       ]);
       const cartId = cart.rows[0].id;
+
+      // Add qty back into stock_quantity table
+      await pool.query(
+        "UPDATE products SET stock_quantity = (stock_quantity + 1) WHERE id = $1",
+        [product_id]
+      );
 
       await pool.query(
         "UPDATE cart_items SET quantity = quantity - 1 WHERE cart_items.cart_id = $1 AND cart_items.product_id = $2 RETURNING *",
@@ -102,9 +127,39 @@ const decreaseItem = async (req, res) => {
 };
 
 const emptyCart = async (req, res) => {
-  const cartId = parseInt(req.params.id);
-  await pool.query("DELETE FROM cart_items WHERE cart_id = $1", [cartId]);
-  res.status(200).json({ message: "Cart emptied" });
+  const userId = req.user.id;
+
+  if (userId) {
+    try {
+      // Get user's cart id
+      const cart = await pool.query("SELECT * FROM carts WHERE user_id = $1", [
+        userId,
+      ]);
+      const cartId = cart.rows[0].id;
+
+      // Retrieve items from the cart that are about to be removed
+      const itemsInCart = await pool.query(
+        "SELECT product_id, quantity FROM cart_items WHERE cart_id = $1",
+        [cartId]
+      );
+
+      // Increment the stock quantity for each item in the products table
+      for (const item of itemsInCart.rows) {
+        const { product_id, quantity } = item;
+        await pool.query(
+          "UPDATE products SET stock_quantity = stock_quantity + $1 WHERE id = $2",
+          [quantity, product_id]
+        );
+      }
+
+      // Delete items from cart
+      await pool.query("DELETE FROM cart_items WHERE cart_id = $1", [cartId]);
+      res.status(200).json({ message: "Cart emptied" });
+    } catch (error) {
+      console.error(error.message);
+      return res.json({ error: error.message });
+    }
+  }
 };
 
 module.exports = {
